+++ shared/src/test/scala/scalus/uplc/ArbitraryInstances.scala
@@ -51,7 +51,7 @@ trait ArbitraryInstances:
                     Stream(
                       B(
                         builtin.ByteString.unsafeFromArray(
-                          b.bytes.slice(0, b.bytes.length / 2)
+                          b.bytes.slice(0, b.length / 2)
                         )
                       )
                     )